Add files via upload
In the world full of data that seldom follows nice theoretical distribution, the Central  Limit Theorem is a beacon of light. It's an important concept to understand when performing any type of data analysis. This blog will explain what Central Limit Theorem is and it's implementation in Python.
Law of Large Numbers
If we would like to find the average income of a person in America. Collecting data for every person in America is impractical, so we draw samples from it. As we increase the sample size from the population, the average income of samples tends to be closer to the average of the income population. This phenomenon is known as The Law of Large Numbers 
As Sample Size increases:
 - Sample becomes more representative of Population
-Samples mean gets closer to the population mean
Central Limit Theorem
We want to figure out the average income of a person in American. Now we take multiple samples and take an average of each sample, then we calculate the average of the average income of the samples. Surprisingly, we find out the number is very close to the average income of the population. If we plot all of these average values of the sample and form a distribution, we figure out that the average value of our samples follow a normal distribution  with the average on a mean value close to the actual population average. This phenomenon is known as Central Limit Theorem.
If the sample size is large enough, distribution of Sample Means approximates a Gaussian Distribution
Mean of samples approximate the Population Mean
We need Random Samples
Distribution of a Sample can be Non-Guassian but Means of Samples will follow Gaussian Distribution
Instead of collecting entire population data, we can gather a subset of data from a population and use the statistics of that sample to draw conclusions about the population.